Match - sonnet 4.6

+++ include/functions.us.h
@@ -484,7 +484,7 @@ s16 func_800C17B4_D0764(u8 arg0, s32 arg1);
 void func_800C14D4_D0484(u8 arg0);
 void func_800C1E24_D0DD4(s16 arg0, s32 arg1, s32 arg2);
 void func_800C1A4C_D09FC(s16 arg0, s32 arg1, s32 arg2);
-void func_800D0614_DF5C4(s16 arg0, s16 arg1, s16 arg2, s32 arg3, s32 arg4, s32 arg5, s32 arg6, s32 arg7);
+void func_800D0614_DF5C4(s16 arg0, s16 arg1, s16 arg2, u16 arg3, u8 arg4, u8 arg5, u8 arg6, s32 arg7);
 f32 func_800C0FD4_CFF84(Vec3f *arg0);
 void func_800DDB18_ECAC8(void);
 u8 func_800DDB60_ECB10(s16 arg0, s16 arg1, s16 arg2, s32 arg3, s32 arg4);
